--- .storybook/preview.js
@@ -1,11 +1,0 @@
-import '../src/lib/tailwind/theme.css'
-
-export const parameters = {
-  actions: { argTypesRegex: '^on[A-Z].*' },
-  controls: {
-    matchers: {
-      color: /(background|color)$/i,
-      date: /Date$/,
-    },
-  },
-}
