+++ .storybook/preview.ts
@@ -1,0 +1,16 @@
+import type { Preview } from '@storybook/react'
+import '../src/lib/tailwind/theme.css'
+
+const preview: Preview = {
+  parameters: {
+    actions: { argTypesRegex: '^on[A-Z].*' },
+    controls: {
+      matchers: {
+        color: /(background|color)$/i,
+        date: /Date$/,
+      },
+    },
+  },
+}
+
+export default preview
