+++ src/lib/components/atoms/at-button/at-button.stories.tsx
@@ -1,11 +1,11 @@
 import React from 'react'
-import { ComponentStory, ComponentMeta } from '@storybook/react'
-
 import { AtButton, AtButtonProps, AT_BUTTON_VARIANT } from '.'
 import { objectValuesToControls } from '../../../storybook-utils'
+import { Meta } from '@storybook/react'
+import { StoryFn } from '@storybook/react'
 
 // More on default export: https://storybook.js.org/docs/react/writing-stories/introduction#default-export
-export default {
+const meta: Meta<typeof AtButton> = {
   title: 'Atoms/AtButton',
   component: AtButton,
   // More on argTypes: https://storybook.js.org/docs/react/api/argtypes
@@ -14,10 +14,11 @@ export default {
     variant: objectValuesToControls(AT_BUTTON_VARIANT),
     onClick: { action: 'clicked' },
   },
-} as ComponentMeta<typeof AtButton>
+}
+export default meta
 
 // More on component templates: https://storybook.js.org/docs/react/writing-stories/introduction#using-args
-const Template: ComponentStory<typeof AtButton> = (args: AtButtonProps) => <AtButton {...args} />
+const Template: StoryFn<typeof AtButton> = (args: AtButtonProps) => <AtButton {...args} />
 
 export const Primary = Template.bind({})
 // More on args: https://storybook.js.org/docs/react/writing-stories/args
